Refactor ExportCommand using DocBlockResolverAwareTrait
