+++ src/php/Console/ConsoleDependencyProvider.php
@@ -8,7 +8,7 @@ use Gacela\Framework\AbstractDependencyProvider;
 use Gacela\Framework\Container\Container;
 use Phel\Build\Infrastructure\Command\CompileCommand;
 use Phel\Formatter\Infrastructure\Command\FormatCommand;
-use Phel\Interop\InteropFacade;
+use Phel\Interop\Infrastructure\Command\ExportCommand;
 use Phel\Run\Infrastructure\Command\TestCommand;
 use Phel\Run\RunFacade;
 
@@ -18,11 +18,10 @@ final class ConsoleDependencyProvider extends AbstractDependencyProvider
 
     public function provideModuleDependencies(Container $container): void
     {
-        $interopFacade = new InteropFacade();
         $runFacade = new RunFacade();
 
         $container->set(self::COMMANDS, static fn () => [
-            $interopFacade->getExportCommand(),
+            new ExportCommand(),
             new FormatCommand(),
             $runFacade->getReplCommand(),
             $runFacade->getRunCommand(),
