+++ src/php/Interop/Infrastructure/Command/ExportCommand.php
@@ -4,96 +4,39 @@ declare(strict_types=1);
 
 namespace Phel\Interop\Infrastructure\Command;
 
-use Phel\Command\CommandFacadeInterface;
+use Gacela\Framework\DocBlockResolverAwareTrait;
 use Phel\Compiler\Domain\Exceptions\CompilerException;
-use Phel\Interop\Domain\DirectoryRemover\DirectoryRemoverInterface;
-use Phel\Interop\Domain\ExportFinder\FunctionsToExportFinderInterface;
-use Phel\Interop\Domain\FileCreator\FileCreatorInterface;
-use Phel\Interop\Domain\Generator\WrapperGeneratorInterface;
-use Phel\Interop\Domain\ReadModel\Wrapper;
-use RuntimeException;
+use Phel\Interop\InteropFacade;
 use Symfony\Component\Console\Command\Command;
 use Symfony\Component\Console\Input\InputInterface;
 use Symfony\Component\Console\Output\OutputInterface;
 use Throwable;
 
+/**
+ * @method InteropFacade getFacade()
+ */
 final class ExportCommand extends Command
 {
-    public const COMMAND_NAME = 'export';
-
-    private DirectoryRemoverInterface $directoryRemover;
-    private WrapperGeneratorInterface $wrapperGenerator;
-    private FunctionsToExportFinderInterface $functionsToExportFinder;
-    private FileCreatorInterface $fileCreator;
-    private CommandFacadeInterface $commandFacade;
-
-    public function __construct(
-        DirectoryRemoverInterface $directoryRemover,
-        WrapperGeneratorInterface $wrapperGenerator,
-        FunctionsToExportFinderInterface $functionsToExportFinder,
-        FileCreatorInterface $fileCreator,
-        CommandFacadeInterface $commandFacade
-    ) {
-        parent::__construct(self::COMMAND_NAME);
-        $this->directoryRemover = $directoryRemover;
-        $this->wrapperGenerator = $wrapperGenerator;
-        $this->functionsToExportFinder = $functionsToExportFinder;
-        $this->fileCreator = $fileCreator;
-        $this->commandFacade = $commandFacade;
-    }
-
-    /**
-     * @return list<Wrapper>
-     */
-    public function generateWrappers(): array
-    {
-        $allFunctionsToExport = $this->functionsToExportFinder->findInPaths();
-        $wrappers = [];
-
-        foreach ($allFunctionsToExport as $ns => $functionsToExport) {
-            $wrappers[] = $this->wrapperGenerator->generateCompiledPhp($ns, $functionsToExport);
-        }
-
-        return $wrappers;
-    }
+    use DocBlockResolverAwareTrait;
 
     protected function configure(): void
     {
-        $this->setDescription('Export all definitions with the meta data `@{:export true}` as PHP classes.');
+        $this->setName('export')
+            ->setDescription('Export all definitions with the meta data `@{:export true}` as PHP classes.');
     }
 
     protected function execute(InputInterface $input, OutputInterface $output): int
     {
         try {
-            $this->createGeneratedWrappers($output);
+            $this->getFacade()->generateExportCode($output);
 
             return self::SUCCESS;
         } catch (CompilerException $e) {
-            $this->commandFacade->writeLocatedException($output, $e->getNestedException(), $e->getCodeSnippet());
+            $this->getFacade()->writeLocatedException($output, $e);
         } catch (Throwable $e) {
-            $this->commandFacade->writeStackTrace($output, $e);
+            $this->getFacade()->writeStackTrace($output, $e);
         }
 
         return self::FAILURE;
-    }
-
-    /**
-     * @throws CompilerException
-     * @throws RuntimeException
-     */
-    private function createGeneratedWrappers(OutputInterface $output): void
-    {
-        $this->directoryRemover->removeDir();
-        $output->writeln('Exported namespaces:');
-        $wrappers = $this->generateWrappers();
-
-        if (empty($wrappers)) {
-            $output->writeln('No functions were found to be exported');
-        }
-
-        foreach ($wrappers as $i => $wrapper) {
-            $this->fileCreator->createFromWrapper($wrapper);
-            $output->writeln(sprintf('  %d) %s', $i + 1, $wrapper->relativeFilenamePath()));
-        }
     }
 }
