+++ src/php/Interop/InteropFacade.php
@@ -5,15 +5,37 @@ declare(strict_types=1);
 namespace Phel\Interop;
 
 use Gacela\Framework\AbstractFacade;
-use Phel\Interop\Infrastructure\Command\ExportCommand;
+use Phel\Compiler\Domain\Exceptions\CompilerException;
+use Symfony\Component\Console\Output\OutputInterface;
+use Throwable;
 
 /**
  * @method InteropFactory getFactory()
  */
 final class InteropFacade extends AbstractFacade implements InteropFacadeInterface
 {
-    public function getExportCommand(): ExportCommand
+    public function writeLocatedException(OutputInterface $output, CompilerException $e): void
     {
-        return $this->getFactory()->createExportCommand();
+        $this->getFactory()
+            ->getCommandFacade()
+            ->writeLocatedException(
+                $output,
+                $e->getNestedException(),
+                $e->getCodeSnippet()
+            );
+    }
+
+    public function writeStackTrace(OutputInterface $output, Throwable $e): void
+    {
+        $this->getFactory()
+            ->getCommandFacade()
+            ->writeStackTrace($output, $e);
+    }
+
+    public function generateExportCode(OutputInterface $output): void
+    {
+        $this->getFactory()
+            ->createExportCodeGenerator()
+            ->generateExportCode($output);
     }
 }
